requirements.txt
psycopg2-binary==2.9.3
Flask==2.1.1